feat: refactor AnimatedCarousel component to enhance functionality and improve code structure
